Fix the Win Condition Bug
Still needs testing but, the bug was caused by in 2 areas.
1. The shape of the start and finish tile was empty shape ,which cause
the connect function to return false (5). To fix this just make the
start and finish shape be their own type.
2. The Cross shape will return the opposite side of where the tile is
connected to the next tile when it should be giving the side it is
connecting to the next tiles.